backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.13.1
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
